+++ Pulse/Views/Templates/TemplatesView.swift
@@ -151,5 +151,6 @@ struct TemplatesView: View {
             Exercise.self,
             WorkoutTemplate.self,
             TemplateExercise.self,
+            TemplateSet.self,
         ], inMemory: true)
 }